+++ system/src/Grav/Console/Gpm/InstallCommand.php
@@ -21,6 +21,7 @@ use Grav\Console\ConsoleCommand;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Question\ConfirmationQuestion;
+use Symfony\Component\Console\Style\SymfonyStyle;
 
 \define('GIT_REGEX', '/http[s]?:\/\/(?:.*@)?(github|bitbucket)(?:.org|.com)\/.*\/(.*)/');
 
@@ -100,6 +101,13 @@ class InstallCommand extends ConsoleCommand
      */
     protected function serve()
     {
+        if (!class_exists(\ZipArchive::class)) {
+            $io = new SymfonyStyle($this->input, $this->output);
+            $io->title('GPM Install');
+            $io->error('php-zip extension needs to be enabled!');
+            exit;
+        }
+
         $this->gpm = new GPM($this->input->getOption('force'));
 
         $this->all_yes = $this->input->getOption('all-yes');