+++ system/src/Grav/Console/Gpm/UpdateCommand.php
@@ -17,6 +17,7 @@ use Symfony\Component\Console\Input\ArrayInput;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Question\ConfirmationQuestion;
+use Symfony\Component\Console\Style\SymfonyStyle;
 
 class UpdateCommand extends ConsoleCommand
 {
@@ -96,6 +97,13 @@ class UpdateCommand extends ConsoleCommand
 
     protected function serve()
     {
+        if (!class_exists(\ZipArchive::class)) {
+            $io = new SymfonyStyle($this->input, $this->output);
+            $io->title('GPM Update');
+            $io->error('php-zip extension needs to be enabled!');
+            exit;
+        }
+
         $this->upgrader = new Upgrader($this->input->getOption('force'));
         $local = $this->upgrader->getLocalVersion();
         $remote = $this->upgrader->getRemoteVersion();